\name{seqbias}
\docType{class}
\alias{seqbias-class}

\title{The seqbias model}

\description{
The \code{seqbias} class maintains a model of the sequencing bias of from an
experiment, which can be saved, loaded, trained, and used to make predictions of
bias.

The class is manipulated with the following functions:
\code{\link{seqbias.fit}}
\code{\link{seqbias.load}}
\code{\link{seqbias.predict}}
\code{\link{seqbias.save}}
}
\author{
    Daniel Jones
    \email{dcjones@cs.washington.edu}
}


