\name{seqbias-package}
\docType{package}
\alias{seqbias-package}
\alias{seqbias}
\title{'seqbias' modeling bias in high-throughput sequencing data}
\description{
    This package implements a model of sequencing bias in high-throughput
    sequencing data using a simple Bayesian network, the structure and
    parameters of which are trained on a set of aligned reads and a reference
    genome sequence.
}
\author{
    Daniel Jones
    \email{dcjones@cs.washington.edu}
}

